requirements.txt
# Data Processing
pandas
numpy

# Vector Search & Embeddings
faiss-cpu
sentence-transformers
torch

# LLM & API
openai
python-dotenv

# Utility
pickle
tqdm